address link2xt comments

+++ src/context.rs
@@ -577,7 +577,6 @@ impl Context {
         let Some((transport_id, param)) = ConfiguredLoginParam::load(self).await? else {
             bail!("Not configured");
         };
-        // A relay advertising its own limit via IMAP METADATA is authoritative.
         let metadata_limit = self
             .metadata
             .read()
@@ -587,10 +586,9 @@ impl Context {
         if let Some(limit) = metadata_limit {
             return Ok(limit);
         }
-        // A few legacy domains (e.g. nauta.cu) need a hard-coded limit.
         if let Some(limit) = crate::provider::legacy_settings_for_addr(&param.addr).max_smtp_rcpt_to
         {
-            return Ok(limit as u32);
+            return Ok(limit);
         }
         Ok(constants::DEFAULT_MAX_SMTP_RCPT_TO)
     }
